fix: Move event handlers into named functions, create custom hook 'useWebSocket' for WebSocket logic
