+++ src/pages/HomePage/index.tsx
@@ -1,18 +1,23 @@
 import IntroSlides from "@components/IntroSlides";
 import WelcomeBlock from "@components/NavigationButton";
 import { type ReactElement, useEffect } from "react";
-import { useNavigate } from "react-router-dom";
+import { useNavigate, type NavigateFunction } from "react-router-dom";
 import "./style.css";
 
-const HomePage = (): ReactElement => {
-    const navigate = useNavigate();
+const nickNameChecker = (navigate: NavigateFunction) => {
+  const nickName = localStorage.getItem("nickName");
 
-    useEffect(() => {
-        const nickName = localStorage.getItem("nickName");
-        if (!nickName) {
-            navigate("/registration");
-        }
-    }, [navigate]);
+  if (!nickName) {
+    navigate("/registration");
+  }
+
+  return;
+}
+
+const HomePage = (): ReactElement => {
+  const navigate = useNavigate();
+
+  useEffect(() => {nickNameChecker(navigate)}, [navigate]);
 
   return (
     <div className="home-page">
