+++ src/pages/MessagePage/index.tsx
@@ -1,71 +1,60 @@
 import Message from "@components/Message";
 import MessageInput from "@components/MessageInput";
+import { useWebSocket } from "@hooks/useWebSocket/useWebSocket.ts";
 import { useState, useRef, useEffect } from "react";
 import { type IMessage } from "@app-types/message";
 import { Link } from "react-router-dom";
 import "./style.css";
 
-function getRandomId() {
-    return (
-        Date.now().toString(36) +
-        Math.random().toString(36).substring(2).toString()
-    );
-}
-
 export default function MessagePage() {
     const [secondUsername, setSecondUsername] = useState<string | null>(null);
     const [messages, setMessages] = useState<IMessage[]>([]);
-
     const [ws, setWs] = useState<WebSocket | null>(null);
-
     const messagesEndRef = useRef<HTMLDivElement>(null);
-
     const username = localStorage.getItem("nickName");
 
     useEffect(() => {
         const socket = new WebSocket("ws://localhost:3001");
 
-        socket.onopen = () => {
-            console.log("Connected to ws");
+        // socket.onopen = () => {
+        //     console.log("Connected to ws");
 
-            socket.send(
-                JSON.stringify({ type: "init", username, id: getRandomId() })
-            );
-        };
+        //     socket.send(
+        //         JSON.stringify({ type: "init", username, id: getRandomId() })
+        //     );
+        // };
 
-        socket.onmessage = (event) => {
-            const data = JSON.parse(event.data);
+        // socket.onmessage = (event) => {
+        //     const data = JSON.parse(event.data);
 
-            console.log("MESSAGE FROM SERVER: ", data);
+        //     console.log("MESSAGE FROM SERVER: ", data);
 
-            // if (data.type === "init" && data.username !== username) {
-            //     setSecondUser(data.username);
-            // }
+        //     if (data.type === "msg") {
+        //         if (
+        //             !secondUsername &&
+        //             data.username !== localStorage.getItem("nickName")
+        //         ) {
+        //             setSecondUsername(data.username);
 
-            if (data.type === "msg") {
-                if (
-                    !secondUsername &&
-                    data.username !== localStorage.getItem("nickName")
-                ) {
-                    setSecondUsername(data.username);
+        //             console.log("SECOND USERNAME CHANGED: ", data.username);
+        //         }
+        //         const newMessage: IMessage = {
+        //             id: Date.now().toString(),
+        //             text: data.text,
+        //             time: new Date().toLocaleTimeString([], {
+        //                 hour: "2-digit",
+        //                 minute: "2-digit",
+        //             }),
+        //             isMine: data.username === localStorage.getItem("nickName"),
+        //             sender: data.username,
+        //         };
+        //         if (localStorage.getItem("nickName") !== data.username) {
+        //             setMessages((prev) => [...prev, newMessage]);
+        //         }
+        //     }
+        // };
 
-                    console.log("SECOND USERNAME CHANGED: ", data.username);
-                }
-                const newMessage: IMessage = {
-                    id: Date.now().toString(),
-                    text: data.text,
-                    time: new Date().toLocaleTimeString([], {
-                        hour: "2-digit",
-                        minute: "2-digit",
-                    }),
-                    isMine: data.username === localStorage.getItem("nickName"),
-                    sender: data.username,
-                };
-                if (localStorage.getItem("nickName") !== data.username) {
-                    setMessages((prev) => [...prev, newMessage]);
-                }
-            }
-        };
+        useWebSocket(username, secondUsername, setSecondUsername, setMessages, socket);
 
         setWs(socket);
