+++ tsconfig.app.json
@@ -29,7 +29,8 @@
             "@pages/*": ["src/pages/*"],
             "@app-types/*": ["src/types/*"],
             "@router/*":["src/router/*"],
-            "@utils/*": ["src/utils/*"]
+            "@utils/*": ["src/utils/*"],
+            "@hooks/*": ["src/hooks/*"]
         }
     },
     "include": ["src"]
