+++ tsconfig.node.json
@@ -28,7 +28,8 @@
             "@pages/*": ["src/pages/*"],
             "@app-types/*": ["src/types/*"],
             "@router/*" : ["src/router/*"],
-            "@utils/*" : ["src/utils/*"]
+            "@utils/*" : ["src/utils/*"],
+            "@hooks/*": ["src/hooks/*"]
         }
     },
     "include": ["vite.config.ts"]
